requirements.txt
requests==2.31.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
openpyxl==3.1.2